requirements.txt
streamlit
python-dotenv
google.generativeai
langchain
PyPDF2
pinecone-client
sumy
nltk